requirements.txt
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
sympy==1.11.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.13
torchsummary==1.5.1